requirements.txt
aiofiles==0.8.0
aiosqlite==0.17.0
async-generator==1.10
asyncua==0.9.94
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.0
cryptography==37.0.2
et-xmlfile==1.1.0
exceptiongroup==1.0.0
h11==0.14.0
idna==3.4
numpy==1.24.3
opcua-client==0.8.4
opcua-modeler==0.5.14
opcua-widgets==0.6.1
openpyxl==3.1.2
outcome==1.2.0
pandas==2.0.1
pycparser==2.21
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.2.0
tzdata==2023.3
urllib3==1.26.12
wsproto==1.2.0
